refactor(doctor): 診断本体を readiness 層へ分離する (#4448)
